old-tests/submission/W3C/domhtml/HTMLIFrameElement09.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-iframe-element.html#dom-iframe-src" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe src="/support/html/sample.html">IFRAME1</iframe>


<script>
var domtest = async_test("The IDL attribute src must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vsrc;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      vsrc = testNode.src;

     assert_uri_equals(document.location.protocol,null,document.location.host,"sample.html",null,null,null,true,vsrc);
    domtest.done();
}));

</script>
<p>The IDL attribute src must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
